requirements.txt
asgiref==3.7.2
Django==4.2.9
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
msoffice2pdf==0.0.8
openpyxl==3.1.2
packaging==23.2
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.9.0
django-cors-headers==4.3.1
